Upvoter
This is a script using the Reddit API that will upvote posts based on
subreddit and keywords.  For example, if anyone posts in 'r/learnpython'
and they say 'please', 'pls', 'or 'plz' they will get an upvote.  The
script will run and then stop for 5 minutes and then run again.  It
tracks posts that have been upvoted in the 'upvoted.txt' file.